requirements.txt
opencv-python
scikit-learn
numpy
pandas
Pillow 
plotly
matplotlib
streamlit
pixellib
webcolors
KMeans
tensorflow

